% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beamer.R
\name{beamer_presentation_ja}
\alias{beamer_presentation_ja}
\title{R Markdown 上で XeLaTeX を使い日本語 beamer スライドを作成するフォーマット}
\usage{
beamer_presentation_ja(
  keep_tex = TRUE,
  keep_md = FALSE,
  theme = "default",
  theme_options = "default",
  fonttheme = "default",
  colortheme = "default",
  toc = FALSE,
  fig_width = 5.03937,
  fig_height = 3.77953,
  fig_crop = TRUE,
  fig_caption = TRUE,
  out.width = "100\%",
  out.height = "100\%",
  extract_media = F,
  highlight = "default",
  code_rownumber = FALSE,
  code_softwrapped = TRUE,
  block_style = c("default", "kframe", "tcolorbox"),
  citation_package = "biblatex",
  citation_options = "default",
  latexmk_emulation = !citation_package == "natbib",
  figurename = "図",
  tablename = "表",
  number_sections = FALSE,
  slide_level = 2,
  incremental = FALSE,
  self_contained = TRUE,
  includes = NULL,
  latex_engine = c("xelatex", "lualatex", "pdflatex", "tectonic"),
  dev = "cairo_pdf",
  df_print = "default",
  template = "default",
  md_extensions = "+raw_tex",
  knitr_options = NULL,
  pandoc_args = NULL
)
}
\arguments{
\item{keep_tex}{Keep the intermediate tex file used in the conversion to PDF}

\item{keep_md}{Keep the markdown file generated by knitting.}

\item{theme}{chracter. beamer テーマ.}

\item{theme_options}{character. テーマオプション. デフォルトはフレームタイトルの下にプログレスバーをつけて, ブロックの背景色を描画するというもの}

\item{fonttheme}{character. フォントテーマ. デフォルトでは数式にローマン体を使う.}

\item{colortheme}{character. 色テーマ.}

\item{toc}{logical. 目次をスライド冒頭に出力するかどうか. examples のように自分で書いたほうが良いかもしれません.}

\item{fig_width}{Default width (in inches) for figures}

\item{fig_height}{Default height (in inches) for figures}

\item{fig_crop}{\code{TRUE} to automatically apply the \code{pdfcrop} utility
(if available) to pdf figures}

\item{fig_caption}{\code{TRUE} to render figures with captions}

\item{highlight}{Syntax highlighting style. Supported styles include
"default", "tango", "pygments", "kate", "monochrome", "espresso",
"zenburn", and "haddock". Pass \code{NULL} to prevent syntax highlighting.}

\item{citation_package}{The LaTeX package to process citations, \code{natbib}
or \code{biblatex}. Use \code{default} if neither package is to be used,
which means citations will be processed via the command
\command{pandoc-citeproc}.}

\item{citation_options}{character. \code{citation_package} のオプション. \code{"default"}, 空の文字列, \code{NULL} などを指定すると特に何もしない. \code{citation_package = "natbib"} を選んだ場合, \code{"default"} は \code{`numbers`} に書き換えられる.}

\item{figurename}{character. 図X の「図」の部分のテキスト.}

\item{tablename}{character. 表Y の「表」の部分のテキスト.}

\item{number_sections}{\code{TRUE} to number section headings}

\item{slide_level}{integer. フレームタイトルに対応する markdown の節レベル. デフォルト: 2. つまり \verb{#} はセクションタイトルで, \verb{##} がフレームタイトルになる.}

\item{incremental}{logical. 箇条書きが順番に現れるやつ. 文字が回転するアニメーション機能はない.
『...遠慮のないマッポ関係者が失笑した。ナムサン！プレゼンテーションにおける典型的なセンスレス文字操作だ。』--- B. ボンド& F. モーゼズ}

\item{self_contained}{logical. tex ファイルのプリアンブルも生成するかどうか. texソースを手動で書き換えたいのでない限り \code{TRUE}.}

\item{includes}{Named list of additional content to include within the
document (typically created using the \code{\link[rmarkdown]{includes}} function).}

\item{latex_engine}{LaTeX engine for producing PDF output. Options are
"pdflatex", "lualatex", and "xelatex".}

\item{dev}{Graphics device to use for figure output (defaults to pdf)}

\item{df_print}{Method to be used for printing data frames. Valid values
include "default", "kable", "tibble", and "paged". The "default" method
uses a corresponding S3 method of \code{print}, typically
\code{print.data.frame}. The "kable" method uses the
\code{\link[knitr:kable]{knitr::kable}} function. The "tibble" method uses
the \pkg{tibble} package to print a summary of the data frame. The "paged"
method creates a paginated HTML table (note that this method is only valid
for formats that produce HTML). In addition to the named methods you can
also pass an arbitrary function to be used for printing data frames. You
can disable the \code{df_print} behavior entirely by setting the option
\code{rmarkdown.df_print} to \code{FALSE}. See
\href{https://bookdown.org/yihui/rmarkdown/html-document.html#data-frame-printing}{Data
frame printing section} in bookdown book for examples.}

\item{template}{Pandoc template to use for rendering. Pass "default" to use
the rmarkdown package default template; pass \code{NULL} to use pandoc's
built-in template; pass a path to use a custom template that you've
created.  See the documentation on
\href{https://pandoc.org/MANUAL.html}{pandoc online documentation} for
details on creating custom templates.}

\item{md_extensions}{Markdown extensions to be added or removed from the
default definition or R Markdown. See the \code{\link[rmarkdown]{rmarkdown_format}} for
additional details.}

\item{pandoc_args}{Additional command line options to pass to pandoc}
}
\value{
\code{rmarkdown_output_format} class
}
\description{
\code{bookdown::\link{beamer_presentation2}} wrapper format for Japanese typesetting with XeLaTeX or LuaLaTeX

XeLaTeX または LuaLaTeX で \code{bookdown::beamer_presentation2} で日本語タイプセットをするためのラッパフォーマット.
}
\details{
beamer presentation format for Japanese typesetting with XeLaTeX or LuaLaTeX

rmarkdownで \LaTeX を使う場合, 日本語を適切に表示するためにいろいろ必要だった調整を済ませたフォーマット
基本的なオプションだけに限定することで簡単にポンチ絵スライドになってしまうことを回避する画期的な機能もあります.

テンプレートファイルも変更しているため, 厳密に同じではないですが, 既存のパッケージで表現するならデフォルトは概ねこうなります:\preformatted{output:
  bookdown::beamer_presentation2:
    latex_engine: xelatex
    theme: default
    citation_package: biblatex
    dev: cairo_pdf
    keep_tex: yes
classoption: aspectratio=43
mainfont: ...
sansfont: ...
monofont: ...
biblatexoption:
  - style=jauthoryear
  - citestyle=numeric
}
}
\seealso{
Other pdf formats: 
\code{\link{pdf_book_ja}()},
\code{\link{pdf_output_base}()}
}
\concept{pdf formats}
