% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beamer_presentation_zxja.R
\name{beamer_presentation_zxja}
\alias{beamer_presentation_zxja}
\title{XeLaTeX + beamer + zxjatype}
\usage{
beamer_presentation_zxja(
  incremental = FALSE,
  theme = "metropolis",
  fig_width = 5.03937,
  fig_height = 3.77953,
  fig_crop = TRUE,
  fig_caption = TRUE,
  theme_options = "progressbar=frametitle,block=fill",
  fonttheme = "professionalfonts",
  colortheme = "default",
  innertheme = "default",
  outertheme = "default",
  keep_tex = TRUE,
  keep_md = FALSE,
  fontsize = "11pt",
  mainfont = "Roboto Slab",
  mainfont_options = NULL,
  sansfont = "Roboto",
  sasnfont_options = NULL,
  monofont = "Ricty Diminished",
  monofont_options = NULL,
  mainfont_ja = "Noto Serif CJK JP",
  mainfont_ja_options = NULL,
  sansfont_ja = "Noto Sans CJK JP",
  sansfont_ja_options = NULL,
  monofont_ja = "Ricty Diminished",
  monofont_ja_options = NULL,
  self_contained = TRUE,
  figurename = "図",
  tablename = "表"
)
}
\arguments{
\item{incremental}{logical. 箇条書きが順番に現れるやつ. 『ナムサン！プレゼンテーションにおける典型的なセンスレス文字操作だ。』--- B. ボンド& F. モーゼズ}

\item{theme}{chracter. beamer テーマ. デフォルトは "metropolis"}

\item{fig_width}{numeric. 画像保存時の幅. 単位インチ. デフォルトはbeamerのデフォルト幅と同じ 5.03937}

\item{fig_height}{numeric. 画像保存時の高さ. 単位インチ. デフォルトはbeamerのデフォルト高と同じ 3.77953}

\item{fig_crop}{\code{TRUE} to automatically apply the \code{pdfcrop} utility
(if available) to pdf figures}

\item{fig_caption}{logical. 画像キャプションの有無. デフォルト: TRUE}

\item{theme_options}{character. テーマオプション. デフォルトはフレームタイトルの下にプログレスバーをつけて, ブロックの背景色を描画するというもの}

\item{fonttheme}{character. フォントテーマ. デフォルトでは数式にローマン体を使う. デフォルト: "professional"}

\item{colortheme}{character. 色テーマ. デフォルト: "default"}

\item{innertheme}{character.  \emph{colortheme} に同じ}

\item{outertheme}{character. \emph{outertheme} に同じ}

\item{keep_tex}{logical. 出力時に .tex ファイルを残すかどうか. デバッグのため TRUEを推奨する. デフォルト: TRUE}

\item{keep_md}{logical. 出力時に .md ファイルを残すかどうか. デフォルト: FALSE}

\item{fontsize}{character. beamer の規定フォントサイズ. デフォルト: "11pt"}

\item{mainfont}{character. セリフフォント名. デフォルト: "Roboto Slab"}

\item{mainfont_options}{character. セリフフォントのオプション. デフォルト: NULL}

\item{sansfont}{character. _font\\emph{serif} に同じ. デフォルト: "Roboto"}

\item{monofont}{character. _font\\emph{serif} に同じ. デフォルト: "Ricty Diminished"}

\item{monofont_options}{character. _font\_sans\\emph{options} に同じ.}

\item{mainfont_ja}{character. 日本語用セリフフォント(明朝体). 詳しくは \code{ZxJatype} のドキュメント参照. デフォルト: Noto Serif CJK JP}

\item{sansfont_ja}{character. _font\_serif\\emph{ja} に同じ. デフォルト: Noto Sans CJK JP}

\item{self_contained}{Whether to generate a full LaTeX document (\code{TRUE})
or just the body of a LaTeX document (\code{FALSE}). Note the LaTeX
document is an intermediate file unless \code{keep_tex = TRUE}.}

\item{figurename}{character. 図X の「図」の部分のテキスト. デフォルト: "図"}

\item{tablename}{character. 表Y の「表」の部分のテキスト. デフォルト: "表"}

\item{sansfont_options}{character. _font\_sans\\emph{options} に同じ.}
}
\description{
XeLaTeX + beamer + zxjatype
}
