% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bookdown_file}
\alias{bookdown_file}
\title{---- copy from bookdown package v0.22 ----
https://github.com/rstudio/bookdown/blob/dfc4f1fb04a7ead9c462e40180f0c3a5dbc212c9/R/utils.R}
\usage{
bookdown_file(...)
}
\description{
---- copy from bookdown package v0.22 ----
https://github.com/rstudio/bookdown/blob/dfc4f1fb04a7ead9c462e40180f0c3a5dbc212c9/R/utils.R
}
