% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccicon.R
\name{get_CC}
\alias{get_CC}
\title{Creative Commons のアイコンを取得する関数}
\usage{
get_CC(
  adapt = c("yes", "no", "inherit"),
  commercial = T,
  size = c("normal", "compact"),
  language = "ja"
)
}
\arguments{
\item{adapt:}{character. \code{yes}, \code{no}, \code{inherit} から選択. 再配布の許諾に関する選択.}

\item{commercial:}{logical. 商用利用を許可するかどうか}

\item{size:}{character. \code{normal}, \code{compact} から選択. 画像のサイズ.}

\item{language:}{character. default: \code{ja}. 声明文の言語を選択.}
}
\value{
character. ハイパーリンク付き画像と評価される Markdown の構文
}
\description{
Creative Commons のアイコン画像を簡単に文書に埋め込むための関数
}
