% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_settings.R
\name{get_default_font_family}
\alias{get_default_font_family}
\title{get OS standard font family for ggplot2 and standard graphic device}
\usage{
get_default_font_family(engine = "xelatex")
}
\arguments{
\item{character.}{LaTeX エンジンの標準フォントと合わせるため}
}
\description{
使用中のOSから標準仕様のフォントファミリ名を取得する
}
