% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.R
\name{gitbook_ja}
\alias{gitbook_ja}
\alias{html_document2_ja}
\title{bookdown::gitbook wrapper for Japanese}
\usage{
gitbook_ja(
  fig_caption = TRUE,
  fig_align = "center",
  fig_width = 7,
  fig_height = 5,
  fig_retina = 2,
  table_css = TRUE,
  toc_depth = 3,
  number_sections = TRUE,
  section_divs = TRUE,
  split_by = c("chapter", "chapter+number", "section", "section+number", "rmd", "none"),
  self_contained = FALSE,
  split_bib = TRUE,
  code_folding = c("none", "show", "hide"),
  code_download = FALSE,
  code_rownumber = TRUE,
  highlight = "default",
  dev = "png",
  dev.args = list(res = 200),
  df_print = "default",
  mathjax = "default",
  template = "default",
  config = list(),
  extra_dependencies = NULL,
  css = NULL,
  includes = NULL,
  keep_md = FALSE,
  lib_dir = "libs",
  md_extensions = NULL,
  pandoc_args = NULL,
  ...
)

html_document2_ja(
  fig_caption = TRUE,
  fig_align = "center",
  fig_width = 7,
  fig_height = 5,
  fig_retina = 2,
  table_css = TRUE,
  toc_depth = 3,
  number_sections = TRUE,
  self_contained = FALSE,
  code_folding = c("none", "show", "hide"),
  code_download = FALSE,
  code_rownumber = TRUE,
  highlight = "default",
  dev = "png",
  dev.args = list(res = 200),
  df_print = "default",
  mathjax = "default",
  template = "default",
  extra_dependencies = NULL,
  css = NULL,
  includes = NULL,
  keep_md = FALSE,
  lib_dir = "libs",
  md_extensions = NULL,
  pandoc_args = NULL,
  ...
)
}
\arguments{
\item{fig_caption}{...
Arguments to be passed to \code{rmarkdown::\link{html_document}()}
(\code{...} not including \code{toc}, and \code{theme}).}

\item{table_css}{\code{TRUE} to load gitbook's default CSS for tables. Choose
\code{FALSE} to unload and use customized CSS (for exmaple, bootstrap) via
the \code{css} option. Default is \code{TRUE}.}

\item{number_sections}{...
Arguments to be passed to \code{rmarkdown::\link{html_document}()}
(\code{...} not including \code{toc}, and \code{theme}).}

\item{split_by}{character. デフォルト: 'chapter'. 'chapter', 'chapter+number', 'section', 'section+number', 'rmd', 'none' から選ぶ. ページを区切る単位.}

\item{self_contained}{...
Arguments to be passed to \code{rmarkdown::\link{html_document}()}
(\code{...} not including \code{toc}, and \code{theme}).}

\item{split_bib}{Whether to split the bibliography onto separate pages where
the citations are actually used.}

\item{template}{Pandoc template to use for rendering. Pass \code{"default"}
to use the bookdown default template; pass a path to use a custom template.
The default template should be sufficient for most use cases. In case you
want to develop a custom template, we highly recommend to start from the
default template:
\url{https://github.com/rstudio/bookdown/blob/master/inst/templates/gitbook.html}.}

\item{config}{A list of configuration options for the gitbook style, such as
the font/theme settings.}

\item{lib_dir}{...
Arguments to be passed to \code{rmarkdown::\link{html_document}()}
(\code{...} not including \code{toc}, and \code{theme}).}

\item{pandoc_args}{...
Arguments to be passed to \code{rmarkdown::\link{html_document}()}
(\code{...} not including \code{toc}, and \code{theme}).}

\item{...}{Other arguments to be passed to \code{base_format}. For
\code{html_book()} and \code{tufte_html_book()}, \code{...} is passed to
\code{html_chapters()}.}

\item{config:}{list. デフォルト: 実際に出力して確認したほうが早い. 詳しくは bookdown のマニュアル.}

\item{dev:}{character. デフォルト: \code{png} しかし将来 \code{ragg_png} にするかもしれない.}

\item{code_rownumber:}{logical. デフォルト: TRUE. 行番号を表示するかどうか}
}
\description{
bookdown::gitbook 日本語版
}
\details{
PDF に比べて大したことはしてないです. というか gitbook/pdf_document でも気にならないかも.
\itemize{
\item \code{gitbook_ja} は \code{bookdown::\link{gitbook}} の日本語版です.
\item \code{html_document2_ja} は \code{bookdown::\link{html_document2}} の日本語版です.
}
}
