% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user-utils.R
\name{include_graphics_svg}
\alias{include_graphics_svg}
\title{PDF でもSVG画像を貼り付けるためのラッパ関数}
\usage{
include_graphics_svg(path, ...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[knitr:include_graphics]{knitr::include_graphics}}
  \describe{
    \item{\code{path}}{A character vector of image paths.}
    \item{\code{auto_pdf}}{Whether to use PDF images automatically when the output
format is LaTeX. If \code{TRUE}, then e.g. \file{foo/bar.png} will be
replaced by \file{foo/bar.pdf} if the latter exists. This can be useful
since normally PDF images are of higher quality than raster images like
PNG, when the output is LaTeX/PDF.}
    \item{\code{dpi}}{DPI (dots per inch) value. Used to calculate the output width (in
inches) of the images. This will be their actual width in pixels, divided
by \code{dpi}. If not provided, the chunk option \code{dpi} is used; if
\code{NA}, the output width will not be calculated.}
    \item{\code{error}}{Whether to signal an error if any files specified in the
\code{path} argument do not exist and are not web resources.}
  }}
}
\description{
PDF でもSVG画像を貼り付けるためのラッパ関数
}
\details{
使用法は \code{knitr::\link{include_graphics}} と基本的に同じ. PDF 出力かつ SVG のときに \code{rsvg} パッケージを使用して画像を PDF に変換する. https://stackoverflow.com/questions/50165404/how-to-make-a-pdf-using-bookdown-including-svg-images に基づいて作成.
}
