% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruby.R
\name{ruby}
\alias{ruby}
\title{HTML/PDF双方でルビを表示する}
\usage{
ruby(word, ruby)
}
\arguments{
\item{word}{character. 単語.}

\item{ruby}{character. ルビとして表示する文字列.}
}
\value{
character. knitr を使い html, tex それぞれのソースに応じて出力を変えるようなコード.
}
\description{
\verb{<ruby>} タグと  \code{pxrubrica.sty} を使いタグを出力
}
\details{
HTMLタグやLaTeXコードだけだとどちらか片方にしかルビが表示されないため, \verb{<ruby>} タグと  \code{pxrubrica.sty} を使い分けどちらにもタグが表示されるようにした. LaTeX のほうは \verb{\\ruby[g]\{...\}\{...\} に対応したルビを出力する.}
}
