% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_settings.R
\name{set_graphics_font}
\alias{set_graphics_font}
\title{set font family for ggplot2 and standard graphic device}
\usage{
set_graphics_font(family)
}
\arguments{
\item{family:}{character. フォントファミリ.}
}
\description{
\code{ggplot2} とグラフィックスのテキストのフォント設定を一括で行う
}
\details{
\code{ggplot2} の \code{theme()} および \code{geom_text()}, \code{geom_label()}, そして \code{ggrepel]} をロードしているなら \code{geom_text_repel()}, \code{geom_label_repel()} のフォントファミリを一括設定する.
}
