% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user-utils.R
\name{setup_tinytex_rmdja}
\alias{setup_tinytex_rmdja}
\title{TeX 環境構築用関数}
\usage{
setup_tinytex_rmdja()
}
\description{
TeX 環境構築用関数
}
\details{
rmdja に必要な Tinytex とかの設定を一括して実行する. tinytex や TeX Live 以外の方法で TeX をインストールしている場合の動作は保証しません.
\enumerate{
\item TeX が未インストールなら tinytex::install_tinytex() でインストール
\item 念の為最低限必要なパッケージをインストール
}
}
