% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texlogo.R
\name{texlogo}
\alias{texlogo}
\title{PDF/HTML でロゴを出す}
\usage{
texlogo(logoname, isrichtext = FALSE)
}
\arguments{
\item{logoname:}{character. BXtexlogo に対応したロゴを出す. 大文字小文字問わずマッチする.}
}
\description{
PDF は \code{BXtexlogo.sty} でロゴを出す.
}
\details{
うろ覚えでもいいように大文字小文字問わない. https://github.com/zr-tex8r/BXtexlogo を参考に
}
