% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_book_ja.R
\name{gitbook_ja}
\alias{gitbook_ja}
\title{bookdown::gotbook wrapper for Japanese}
\usage{
gitbook_ja(
  fig_caption = TRUE,
  fig_align = "center",
  fig_width = 7,
  fig_height = 5,
  fig_retina = 2,
  table_css = TRUE,
  toc = TRUE,
  toc_depth = 3,
  toc_float = TRUE,
  number_sections = TRUE,
  section_divs = TRUE,
  split_by = "chapter",
  self_contained = FALSE,
  split_bib = TRUE,
  code_folding = "none",
  code_download = FALSE,
  theme = "default",
  highlight = "default",
  dev = "png",
  dev.args = list(res = 200),
  df_print = "default",
  mathjax = "default",
  template = "default",
  config = list(),
  extra_dependencies = NULL,
  css = NULL,
  includes = NULL,
  keep_md = FALSE,
  lib_dir = "libs",
  md_extensions = NULL,
  pandoc_args = NULL,
  ...
)
}
\arguments{
\item{split_by}{character. デフォルト: 'chapter'. 'chapter', 'chapter+number', 'section', 'section+number', 'rmd', 'none' から選ぶ. ページを区切る単位.}

\item{config}{list. デフォルト: 実際に出力して確認したほうが早い. 詳しくは bookdown のマニュアル.}
}
\description{
bookdown::gotbook wrapper for Japanese
}
