% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_book_ja.R
\name{pdf_book_ja}
\alias{pdf_book_ja}
\title{\code{rmarkdown} + \code{bookdown} で簡単に日本語文書を作るためのプリセットフォーマット}
\usage{
pdf_book_ja(
  toc = TRUE,
  toc_unnumbered = TRUE,
  toc_depth = 3,
  toc_bib = FALSE,
  toc_appendix = FALSE,
  number_sections = TRUE,
  fig_align = "center",
  fig_caption = TRUE,
  fig_crop = TRUE,
  fig_height = 4.5,
  fig_width = 6.5,
  dev = "cairo_pdf",
  dev_args = NULL,
  out_width = "100\%",
  out_height = "100\%",
  out_extra = "keepaspectratio",
  df_print = "default",
  quote_footer = NULL,
  highlight = "default",
  highlight_bw = FALSE,
  code_rownumber = TRUE,
  tombow = FALSE,
  add_folio = FALSE,
  template = "default",
  keep_tex = TRUE,
  keep_md = TRUE,
  latex_engine = "xelatex",
  citation_package = "natbib",
  citation_options = "default",
  includes = NULL,
  md_extensions = NULL,
  output_extensions = NULL,
  pandoc_args = NULL,
  extra_dependencies = NULL,
  base_format = rmarkdown::pdf_document,
  ...
)
}
\arguments{
\item{toc}{See
\code{rmarkdown::\link{pdf_document}}, or the documentation of the
\code{base_format} function.}

\item{toc_unnumbered}{Whether to add unnumberred headers to the table of
contents.}

\item{toc_bib}{Whether to add the bibliography section to the table of
contents.}

\item{toc_appendix}{Whether to add the appendix to the table of contents.}

\item{number_sections}{See
\code{rmarkdown::\link{pdf_document}}, or the documentation of the
\code{base_format} function.}

\item{fig_caption}{See
\code{rmarkdown::\link{pdf_document}}, or the documentation of the
\code{base_format} function.}

\item{quote_footer}{If a character vector of length 2 and the quote footer
starts with three dashes (\samp{---}), \code{quote_footer[1]} will be
prepended to the footer, and \code{quote_footer[2]} will be appended; if
\code{NULL}, the quote footer will not be processed.}

\item{highlight_bw}{Whether to convert colors for syntax highlighting to
black-and-white (grayscale).}

\item{code_rownumber}{デフォルト: TRUE. logical. コードセルに行番号を表示するかどうか.}

\item{tombow}{logical. デフォルト: FALSE. 製本時に必要なトンボ (trim markers) を付けるかどうか. トンボは \code{gentombow.sty} で作成される.}

\item{add_folio}{logica. デフォルト: FALSE. 製本時に全ページにノンブルが必要な場合があるらしいので全ページに表示したい時に.}

\item{pandoc_args}{See
\code{rmarkdown::\link{pdf_document}}, or the documentation of the
\code{base_format} function.}

\item{base_format}{An output format function to be used as the base format.}

\item{...}{Other arguments to be passed to \code{base_format}.}
}
\value{
rmarkdown_output_format
}
\description{
\code{rmarkdown} + \code{bookdown} で PDF文書をビルドする場合, 日本語を適切に表示するためにいろいろ必要だった調整を済ませたフォーマット. 基本的に YAML フロントマター (+ \verb{_ouput.yml}) や knitr チャンクオプションで設定できることをデフォルト値として埋め込んだだけ. \code{index.Rmd}
}
\details{
bookdown::pdf_book wrapper for Japanese typesetting with XeLaTeX/LuaLaTeX
}
