% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf.R
\name{pdf_book_ja}
\alias{pdf_book_ja}
\alias{pdf_document2_ja}
\alias{pdf_document2_platex}
\title{\code{rmarkdown} + \code{bookdown} で簡単に日本語文書を作るためのプリセットフォーマット}
\usage{
pdf_book_ja(
  toc = TRUE,
  toc_unnumbered = TRUE,
  toc_depth = 3,
  toc_bib = FALSE,
  toc_appendix = FALSE,
  number_sections = TRUE,
  fig_align = "center",
  fig_caption = TRUE,
  fig_crop = TRUE,
  fig_height = 4.5,
  fig_width = 6.5,
  dev = "cairo_pdf",
  dev_args = NULL,
  out_width = "100\%",
  out_height = "100\%",
  out_extra = "keepaspectratio",
  df_print = "default",
  quote_footer = NULL,
  highlight = "default",
  highlight_bw = FALSE,
  code_rownumber = TRUE,
  code_softwrapped = TRUE,
  block_style = c("default", "kframe", "tcolorbox"),
  tombow = FALSE,
  add_folio = FALSE,
  template = "default",
  keep_tex = TRUE,
  keep_md = TRUE,
  latex_engine = c("xelatex", "lualatex", "tectonic"),
  citation_package = "biblatex",
  citation_options = "default",
  latexmk_emulation = !citation_package == "natbib",
  extract_media = FALSE,
  includes = NULL,
  md_extensions = NULL,
  output_extensions = NULL,
  pandoc_args = NULL,
  extra_dependencies = NULL,
  ...
)

pdf_document2_ja(
  toc = FALSE,
  toc_unnumbered = TRUE,
  toc_depth = 3,
  toc_bib = FALSE,
  toc_appendix = FALSE,
  number_sections = TRUE,
  fig_align = "center",
  fig_caption = TRUE,
  fig_crop = TRUE,
  fig_height = 4.5,
  fig_width = 6.5,
  dev = "cairo_pdf",
  dev_args = NULL,
  out_width = "100\%",
  out_height = "100\%",
  out_extra = "keepaspectratio",
  df_print = "default",
  quote_footer = NULL,
  highlight = "default",
  highlight_bw = FALSE,
  code_rownumber = TRUE,
  code_softwrapped = TRUE,
  block_style = c("default", "kframe", "tcolorbox"),
  tombow = FALSE,
  add_folio = FALSE,
  template = "default",
  keep_tex = TRUE,
  keep_md = TRUE,
  latex_engine = c("xelatex", "lualatex", "tectonic"),
  citation_package = "biblatex",
  citation_options = "default",
  latexmk_emulation = !citation_package == "natbib",
  extract_media = FALSE,
  includes = NULL,
  md_extensions = NULL,
  output_extensions = NULL,
  pandoc_args = NULL,
  extra_dependencies = NULL,
  ...
)

pdf_document2_platex(platex = c("uplatex", "platex", "none"), ...)
}
\arguments{
\item{toc}{See
\code{rmarkdown::\link{pdf_document}}, or the documentation of the
\code{base_format} function.}

\item{toc_unnumbered}{Whether to add unnumberred headers to the table of
contents.}

\item{toc_bib}{Whether to add the bibliography section to the table of
contents.}

\item{toc_appendix}{Whether to add the appendix to the table of contents.}

\item{number_sections}{See
\code{rmarkdown::\link{pdf_document}}, or the documentation of the
\code{base_format} function.}

\item{fig_caption}{See
\code{rmarkdown::\link{pdf_document}}, or the documentation of the
\code{base_format} function.}

\item{quote_footer}{If a character vector of length 2 and the quote footer
starts with three dashes (\samp{---}), \code{quote_footer[1]} will be
prepended to the footer, and \code{quote_footer[2]} will be appended; if
\code{NULL}, the quote footer will not be processed.}

\item{highlight_bw}{Whether to convert colors for syntax highlighting to
black-and-white (grayscale).}

\item{code_rownumber}{logical. コードセルに行番号を表示するかどうか.}

\item{code_softwrapped}{logical チャンク内のコードを自動折り返しするかどうか. YAML メタデータ \code{code-softwrapped} でも可.}

\item{block_style}{character. \code{block}/\code{block2} チャンクや Fenced Div. 用のスタイル定義. \code{default}, \code{kframe}, \code{tcolorbox} に対応. \code{default} は \code{kframe} と同じ. YAML メタデータに \code{block-style: tcolorbox}, \code{kframe: true} などと書くこともできる. \code{awesomebox.sty} にも近日対応予定.}

\item{tombow}{logical. 製本時に必要なトンボ (trim markers) を付けるかどうか. トンボは \code{gentombow.sty} で作成される.}

\item{add_folio}{logical. 製本時に全ページにノンブルが必要な場合があるらしいので全ページに表示したい時に.}

\item{citation_options}{character. \code{citation_package} のオプション.}

\item{latexmk_emulation}{logical. パッケージオプション \code{tinytex.latexmk.emulation} に連動する. デフォルトでは, 文献引用エンジンを natbib にしたときのみ \code{FALSE}, それ以外は \code{TRUE}. これは \code{tinytex} が (u)pBibTeX に対応していないため. どうしても BibTeX を使いたい場合以外は操作する必要のない不要なオプションですが, 日本語を含む文書を作成する限りそのような場面はないと思われます.}

\item{extract_media}{logical. markdown 構文の画像貼り付けに URL が使用されていた場合, ダウンロードするかどうか. Pandoc の \verb{--extract-media .} に対応. TODO: 現時点では　TRUE にすると余計な画像ファイルが生成される副作用があります. これは rmarkdown にもある不具合です.}

\item{pandoc_args}{See
\code{rmarkdown::\link{pdf_document}}, or the documentation of the
\code{base_format} function.}

\item{...}{Other arguments to be passed to \code{base_format}.}

\item{platex}{character. platex を使う場合のエンジン. TODO: てか UTF8 以外で動かせるの?}
}
\value{
rmarkdown_output_format
}
\description{
bookdown::pdf_book wrapper format for Japanese typesetting with XeLaTeX or LuaLaTeX /\code{bookdown} で PDF文書をビルドする場合, 日本語を適切に表示するためにいろいろ必要だった調整を済ませたラッパフォーマット.
}
\details{
基本的に YAML フロントマター (+ \verb{_ouput.yml}) や knitr チャンクオプションで設定できることをデフォルト値として埋め込んだだけです.
\itemize{
\item pdf_book_ja は \code{\link{pdf_book}} の日本語版です
\item pdf_document2_ja は \code{\link{pdf_document2}} の日本語版です
\item pdf_document2_platex はさらに uplatex によるコンパイルを想定したバージョンです.
\code{pdf_book_ja()} は \code{index.Rmd} を必ずルートディレクトリに用意してください. \code{pdf_book_ja} は \strong{bookdown} パッケージの \code{\link{pdf_book}} と同様に, Knit では書籍をビルドできません. 詳細は \strong{bookdown} の解説を参考にしてください. それ以外は Knit でビルドできます.
}

テンプレートファイルも変更しているため, 厳密に同じではないですが, 既存のパッケージで表現するなら book の場合はデフォルトは概ねこうなります:\preformatted{output:
  bookdown::pdf_book:
    latex_engine: xelatex
    citation_package: biblatex
    dev: cairo_pdf
    keep_tex: yes
documentclass: bxjsbook
mainfont: ...
sansfont: ...
monofont: ...
biblatexoption:
  - style=jauthoryear
}

加えて knitr のチャンクオプションやフックもいろいろ変更していますが, 基本的にはストレスにならないような動的な設定にしたつもりです. (例: テキストの自動折返し, 文字化けしないようなグラフィックデバイスの選択, 「図」「表」などの頻出キーワードのローカライズなど.)

\code{rmarkdown::\link{pdf_document}} と \code{bookdown::\link{pdf_document2}} の違いは相互参照をサポートしているかどうかです. 技術文書や論文では相互参照は良く使うので後者をベースに日本語文書向けの設定を内蔵しました.

TODO: _platex は 動作確認すらしてません. たぶんまだ動かない
}
\seealso{
Other pdf formats: 
\code{\link{beamer_presentation_ja}()}

Other pdf formats: 
\code{\link{beamer_presentation_ja}()}
}
\concept{pdf formats}
